ai_requirements.txt
# AI Sales Intelligence Agent Dependencies
# Python >= 3.9 required
#
# This file contains specific dependencies for the AI sales intelligence agent
# that processes gym leads using LangChain, Groq API, and Playwright.

# ==============================================================================
# CORE AI/LLM DEPENDENCIES
# ==============================================================================

# LangChain framework for AI agent orchestration
langchain-core==0.3.76
langchain-groq==0.3.8

# Groq API client (auto-installed with langchain-groq)
groq==0.31.1

# ==============================================================================
# WEB AUTOMATION & SCRAPING
# ==============================================================================

# Playwright for browser automation and JavaScript rendering
playwright==1.55.0

# ==============================================================================
# DATA PROCESSING
# ==============================================================================

# Pandas for CSV processing and data manipulation
pandas==2.3.2

# ==============================================================================
# UTILITIES & CLI
# ==============================================================================

# Environment variable management
python-dotenv==1.0.0

# Progress bars for CLI feedback
tqdm==4.67.1

# Note: argparse is built into Python and does not need to be installed

# ==============================================================================
# INSTALLATION INSTRUCTIONS
# ==============================================================================
#
# 1. Install Python packages:
#    pip install -r ai_requirements.txt
#
# 2. Install Playwright browsers:
#    playwright install
#
# 3. Set up environment variables:
#    cp .env.example .env
#    # Edit .env and add your GROQ_API_KEY
#
# ==============================================================================